Included String Validation in main validation method
will still be fast enough

Future improvements to validation method might be to include a check of maxium number of allowed fields for a class. May be an edge case but still important. If you create a class with more fields than the stack size , then validation will throw exception :P